requirements.txt
google-genai>=1.0.0
streamlit>=1.30.0